requirements.txt
python-dotenv~=1.0.0
PyYAML~=6.0.1
requests==2.32.3
SQLAlchemy~=2.0.30
psycopg2~=2.9.9
numpy==1.26.4
pandas==2.0.3
python-docx==1.1.2
pillow==10.4.0
boto3==1.35.13
openpyxl==3.1.2
urllib3==2.2.2
XlsxWriter==3.2.0
wcwidth==0.2.13
smbprotocol==1.14.0
tabulate==0.9.0